Split pincode, setcode
