+++ src/pages/NoPage.tsx
@@ -1,8 +1,9 @@
+import notfound from "../assets/img/404.avif"
 const NoPage = () => {
   return (
     <>
       <div className="flex justify-center items-center">
-        <img className="w-full h-full" src="/src/img/404.avif" alt="" />
+        <img className="w-full h-full" src={notfound} alt="notfoundIMG" />
       </div>
     </>
   );
